Completed Object level permisions implementation in the view
#30

+++ mavfc/experiment/utils.py
@@ -25,7 +25,6 @@ class ObjectCreateMixin:
     
     @method_decorator(login_required)
     def get(self, request):
-        # if not request.user.is_staff: return HttpResponseForbidden()
         return render(
             request,
             self.template_name,
@@ -38,7 +37,6 @@ class ObjectCreateMixin:
     
     @method_decorator(login_required)
     def post(self, request):
-        # if not request.user.is_staff: return HttpResponseForbidden()
         bound_form = self.form_class(request.POST)
         if bound_form.is_valid():
             new_obj = bound_form.save()
@@ -65,8 +63,8 @@ class ObjectUpdateMixin:
     
     @method_decorator(login_required)
     def get(self, request, pk):
-        # if not request.user.is_staff: return HttpResponseForbidden()
         obj = get_object_or_404(self.model, pk=pk)
+        if not obj.user_cud_authorized(request.user): return HttpResponseForbidden()
         return render(
             request,
             self.template_name,
@@ -78,8 +76,8 @@ class ObjectUpdateMixin:
     
     @method_decorator(login_required)
     def post(self, request, pk):
-        # if not request.user.is_staff: return HttpResponseForbidden()
         obj = get_object_or_404(self.model, pk=pk)
+        if not obj.user_cud_authorized(request.user): return HttpResponseForbidden()
         bound_form = self.form_class(request.POST, instance=obj)
         if bound_form.is_valid():
             new_obj = bound_form.save()
@@ -105,8 +103,8 @@ class ObjectDeleteMixin:
     
     @method_decorator(login_required)
     def get(self, request, pk):
-        # if not request.user.is_staff: return HttpResponseForbidden()
         obj = get_object_or_404(self.model, pk=pk)
+        if not obj.user_cud_authorized(request.user): return HttpResponseForbidden()
         return render(
             request,
             self.template_name,
@@ -117,8 +115,8 @@ class ObjectDeleteMixin:
     
     @method_decorator(login_required)
     def post(self, request, pk):
-        # if not request.user.is_staff: return HttpResponseForbidden()
         obj = get_object_or_404(self.model, pk=pk)
+        if not obj.user_cud_authorized(request.user): return HttpResponseForbidden()
         obj.delete()
         success(request, self.model_name+' was successfully deleted.')
         return HttpResponseRedirect(self.success_url)
